Filter for erasing items
