+++ src/components/ItemRegistry/RegistryList.js
@@ -20,6 +20,7 @@ class RegistryList extends Component {
             <span> {el.quantity ? el.quantity : "~"} kg</span>{" "}
             <strong>Notes:</strong>
             <span> {el.notes ? el.notes : "~"}</span>
+            <button>Edit Item</button>
             <button
               onClick={() => {
                 removeItem(idx);
@@ -35,34 +36,3 @@ class RegistryList extends Component {
 }
 
 export default RegistryList;
-// eslint-disable-next-line
-// MOSTRAR INVENTARIO EN FORMA DE TABLA:
-{
-  /* <table>
-        <tbody>
-          <tr>
-            <th>Index:</th>
-            <th>Product:</th>
-            <th>Quality:</th>
-            <th>Quantity:</th>
-            <th>Notes:</th>
-          </tr>
-          {items.map((el, idx) => (
-            <tr key={idx}>
-              <td>{idx + 1} </td>
-              <td> {el.product ? el.product : "~"}</td>{" "}
-              <td> {el.quality ? el.quality : "~"}</td>{" "}
-              <td> {el.quantity ? el.quantity : "~"} kg</td>{" "}
-              <td> {el.notes ? el.notes : "~"}</td>
-              <button
-                onClick={() => {
-                  removeItem(idx);
-                }}
-              >
-                Erase Item
-              </button>
-            </tr>
-          ))}
-        </tbody>
-      </table> */
-}
